fix: move sidebar context/constants to lib/ to fix feature-isolation
useSidebar and width constants were in features/auth/ but imported by
features/entries/, violating the feature-isolation dep-cruiser rule.
Move to lib/sidebar-context.tsx (cross-cutting utility).
